fix(windows): add MSVC compatibility for Windows builds
- Add MSVC-compatible atomic operations in rt_platform.h (pointer load,
  store, exchange, compare-exchange, thread fence using _mm_mfence)
- Rewrite rt_pool.c to use rt_platform.h instead of C11 <stdatomic.h>
- Add overflow detection templates for MSVC in OpHandlerUtils.hpp,
  IntOpSupport.hpp, ConstFold.cpp, and SCCP.cpp
- Implement Windows clipboard functions in vgfx_platform_win32.c
- Add Windows file dialog implementation in vg_filedialog.c
- Fix POSIX function mappings (strdup, strtok_r, strcasecmp) for Windows
- Fix class/struct forward declaration mismatch in OopLoweringContext.hpp
- Add VG_UNUSED macro for MSVC compatibility in vg_codeeditor.c
- Fix std::tolower int-to-char conversion warnings in config.cpp
- Fix variable shadowing in break_spec.cpp and session.cpp
- Add MSVC constexpr step limit flag for RuntimeSignatures.cpp
- Add rt_platform.h includes to runtime files using atomics
- Fix Windows header ordering in rt_guid.c and rt_input_pad.c
- Add mode_t definition for Windows in rt_file_io.c

+++ src/tests/unit/test_codegen_x86_64_spiller.cpp
@@ -68,7 +68,7 @@ TEST(Spiller, SpillsActiveValue)
     alloc.hasPhys = true;
     alloc.phys = PhysReg::RAX;
     AllocationResult result{};
-    result.vregToPhys.emplace(7, PhysReg::RAX);
+    result.vregToPhys.emplace(static_cast<uint16_t>(7), PhysReg::RAX);
 
     std::deque<PhysReg> pool{};
     std::vector<MInstr> prefix{};
